backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.0
Pillow==11.3.0
gunicorn==20.1.0
python-dotenv==1.0.0
PyJWT==2.8.0